base: added CommandLineParser TCLAP
Added the Templatized C++ Command Line Parser.
Homepage: http://tclap.sourceforge.net

It consists of several header files.
I put all but the main one in detail/tclap
The main header "CmdLine.h" is currently in detail (maybe move to base?)
Had to patch the includes in all tclap files because of this. but now it works.