CI: Disable gcov output
The extensive output of gcov issued by codecov is very verbose and not
needed most of the time. Recently the Travis build failed because
suddenly the output in the log more than quadrupled. Disable this ouput
to keep the logs cleaner and prevent failing builds due to extensive
gcov reports.