CI: Use old codecov bash script
It seems codecov changed something in the bash script retrieved from
codecov.io/bash which is not reflected in their official Github
repository. Examining the Travis logs the version (which is probably the
short form of the commit hash) changed. Add this version known to work
explicitely to the CI directory to (hopefully) fix the codecov coverage.