CI: Fix broken coverage reports by Travis due to extensive gcov output
Disabling gcov output is ignored for C/C++ projects and only used for
Obj-C. PR created to fix this.

The problem why the build, or rather the creation of the coverage
reports, failed initially was a change to the codecov bash script as
well as the (already then) quite extensive output from gcov. The option
-X gcovout has been ignored and thus reverting the change scanning the
directories done by codecov "fixed" the problem. The real problem is
just the large output since all C++ headers from gcc with inline methods
(most of them or even all) create gcov files which are scanned and
contribute to the terminal output.

Since gcovout disabling is ignored, the fix is to change the gcov
executable to "gcov > /dev/null" to pipe all the non-stderr output to
/dev/null. Maybe change this once the PR fixing the gcovout option is
merged.